# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "EPC-Intel",
                 "version":  "0.1.0"
             },
    "paths":  {
                  "/api/v1/auth/register":  {
                                                "post":  {
                                                             "tags":  [
                                                                          "auth"
                                                                      ],
                                                             "summary":  "Register",
                                                             "operationId":  "register_api_v1_auth_register_post",
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/UserCreate"
                                                                                                                                     }
                                                                                                                      }
                                                                                             },
                                                                                 "required":  true
                                                                             },
                                                             "responses":  {
                                                                               "201":  {
                                                                                           "description":  "Successful Response",
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/UserResponse"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       }
                                                                                       },
                                                                               "422":  {
                                                                                           "description":  "Validation Error",
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         }
                                            },
                  "/api/v1/auth/login":  {
                                             "post":  {
                                                          "tags":  [
                                                                       "auth"
                                                                   ],
                                                          "summary":  "Login",
                                                          "operationId":  "login_api_v1_auth_login_post",
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/UserLogin"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Successful Response",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {

                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                            "422":  {
                                                                                        "description":  "Validation Error",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      }
                                         },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "UserCreate":  {
                                                          "properties":  {
                                                                             "name":  {
                                                                                          "type":  "string",
                                                                                          "title":  "Name"
                                                                                      },
                                                                             "email":  {
                                                                                           "type":  "string",
                                                                                           "format":  "email",
                                                                                           "title":  "Email"
                                                                                       },
                                                                             "password":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Password"
                                                                                          },
                                                                             "role":  {
                                                                                          "type":  "string",
                                                                                          "title":  "Role"
                                                                                      }
                                                                         },
                                                          "type":  "object",
                                                          "required":  [
                                                                           "name",
                                                                           "email",
                                                                           "password",
                                                                           "role"
                                                                       ],
                                                          "title":  "UserCreate"
                                                      },
                                       "UserLogin":  {
                                                         "properties":  {
                                                                            "email":  {
                                                                                          "type":  "string",
                                                                                          "format":  "email",
                                                                                          "title":  "Email"
                                                                                      },
                                                                            "password":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Password"
                                                                                         }
                                                                        },
                                                         "type":  "object",
                                                         "required":  [
                                                                          "email",
                                                                          "password"
                                                                      ],
                                                         "title":  "UserLogin"
                                                     },
                                       "UserResponse":  {
                                                            "properties":  {
                                                                               "id":  {
                                                                                          "type":  "string",
                                                                                          "title":  "Id"
                                                                                      },
                                                                               "email":  {
                                                                                             "type":  "string",
                                                                                             "format":  "email",
                                                                                             "title":  "Email"
                                                                                         },
                                                                               "name":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Name"
                                                                                        },
                                                                               "role":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Role"
                                                                                        }
                                                                           },
                                                            "type":  "object",
                                                            "required":  [
                                                                             "id",
                                                                             "email",
                                                                             "name",
                                                                             "role"
                                                                         ],
                                                            "title":  "UserResponse"
                                                        },
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           },
                                                                                  "input":  {
                                                                                                "title":  "Input"
                                                                                            },
                                                                                  "ctx":  {
                                                                                              "type":  "object",
                                                                                              "title":  "Context"
                                                                                          }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}